move puppeteer test codes
